// shop/src/lib/api-client/models/types_BanditPolicy.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type types_BanditPolicy = {
    active?: boolean;
    blend_alpha?: number;
    blend_beta?: number;
    blend_gamma?: number;
    brand_cap?: number;
    category_cap?: number;
    co_vis_window_days?: number;
    half_life_days?: number;
    mmr_lambda?: number;
    name?: string;
    notes?: string;
    policy_id?: string;
    popularity_fanout?: number;
    profile_boost?: number;
    rule_exclude_events?: boolean;
};